requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
idna==3.11
paystackapi==2.1.3
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
shortuuid==1.0.13
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.1
